Style rolerolldown (#126)
* Updated border and font styling, RoleDropDown

* updated spacing RoleDropDown

* Updated styling and logic to show RoleDropDown

* Removed border color so it uses page color
