+++ frontend/app/components/RoleDropDown.tsx
@@ -1,11 +1,13 @@
+import { useState } from "react";
 import {
-  Content,
   SanityImageCrop,
   SanityImageHotspot,
   internalGroqTypeReferenceTo,
 } from "sanity/types";
+import { createTexts, useTranslation } from "~/utils/i18n";
 import urlFor from "~/utils/imageUrlBuilder";
-import PortableTextComponent from "./PortableTextComponent";
+import ArrowUp from "/arrow-up.svg";
+import ArrowDown from "/arrow-down.svg";
 
 interface RoleDropDownProps {
   roleGroups: Array<{
@@ -31,29 +33,65 @@ interface RoleDropDownProps {
 }
 
 export default function RoleDropDown({ roleGroups }: RoleDropDownProps) {
+  const { t } = useTranslation();
+  const [openRole, setOpenRole] = useState(false);
   return (
-    <div className="m-2 lg:w-1/3">
-      {roleGroups?.map((roleGroup, index) => (
-        <div key={index} className="w-fit m-4">
-          <h3 className="text-xl">{roleGroup.name}</h3>
-          <div>
-            {roleGroup.roles?.map((role, index) => (
-              <div key={index} className="grid grid-flow-col w-fit gap-6">
-                <img
-                  src={urlFor(role.image?.asset?._ref ?? "")}
-                  alt={role.image?.alt ?? ""}
-                  className="w-36 h-36 object-cover"
-                />
-                <div>
-                  <h4 className="text-lg">{role.occupation}</h4>
-                  <h5>{role.name}</h5>
-                  <span>{role.text}</span>
-                </div>
+    <div className="border m-4">
+      <button
+        className="w-80 h-16 py-4 px-6 grid grid-flow-col"
+        onClick={() => setOpenRole(!openRole)}
+      >
+        <span className="self-center justify-self-start text-xl">
+          {t(texts.roleDropDown)}{" "}
+        </span>
+        <img
+          className="w-6 h-6 self-center justify-self-end"
+          src={openRole ? ArrowUp : ArrowDown}
+          alt={
+            openRole ? t(texts.roleDropDownAltUp) : t(texts.roleDropDownAltDown)
+          }
+        />
+      </button>
+      {openRole && (
+        <div className="w-80">
+          {roleGroups?.map((roleGroup, index) => (
+            <div key={index} className="m-4 mb-10">
+              <h3 className="text-base font-semibold">{roleGroup.name}</h3>
+              <div>
+                {roleGroup.roles?.map((role, index) => (
+                  <div key={index} className="flex flex-row mt-8 w-fit gap-6">
+                    <img
+                      src={urlFor(role.image?.asset?._ref ?? "")}
+                      alt={role.image?.alt ?? ""}
+                      className="w-36 h-36 object-cover"
+                    />
+                    <div>
+                      <h4 className="text-lg mb-2">{role.occupation}</h4>
+                      <h5 className="text-base mb-2">{role.name}</h5>
+                      <span>{role.text}</span>
+                    </div>
+                  </div>
+                ))}
               </div>
-            ))}
-          </div>
+            </div>
+          ))}
         </div>
-      ))}
+      )}
     </div>
   );
 }
+
+const texts = createTexts({
+  roleDropDown: {
+    nb: "Medvirkende",
+    en: "Participants",
+  },
+  roleDropDownAltUp: {
+    nb: "Pil opp",
+    en: "Arrow Up",
+  },
+  roleDropDownAltDown: {
+    nb: "Pil ned",
+    en: "Arrow Down",
+  },
+});
